mas/calender-mcp/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
uvicorn[standard]
fastapi
fastmcp
fastapi-mcp
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib